\name{METRIC_CENTER}
\alias{METRIC_CENTER}
\alias{METRIC_SPREAD}
\alias{METRIC_SHIFT}
\alias{METRIC_RATIO}
\alias{METRIC_DISPARITY}
\title{Metric Constants}
\description{
String constants for the metric names used in \code{\link{Threshold}} and
\code{\link{compare1}}/\code{\link{compare2}}.

\describe{
\item{\code{METRIC_CENTER}}{\code{"center"}}
\item{\code{METRIC_SPREAD}}{\code{"spread"}}
\item{\code{METRIC_SHIFT}}{\code{"shift"}}
\item{\code{METRIC_RATIO}}{\code{"ratio"}}
\item{\code{METRIC_DISPARITY}}{\code{"disparity"}}
}
}
\usage{
METRIC_CENTER
METRIC_SPREAD
METRIC_SHIFT
METRIC_RATIO
METRIC_DISPARITY
}
\seealso{
\code{\link{Threshold}}, \code{\link{compare1}}, \code{\link{compare2}}
}
