\name{VERDICT_LESS}
\alias{VERDICT_LESS}
\alias{VERDICT_GREATER}
\alias{VERDICT_INCONCLUSIVE}
\title{Verdict Constants}
\description{
String constants for the verdict values returned in \code{\link{Projection}} results.

\describe{
\item{\code{VERDICT_LESS}}{\code{"less"}}
\item{\code{VERDICT_GREATER}}{\code{"greater"}}
\item{\code{VERDICT_INCONCLUSIVE}}{\code{"inconclusive"}}
}
}
\usage{
VERDICT_LESS
VERDICT_GREATER
VERDICT_INCONCLUSIVE
}
\seealso{
\code{\link{Projection}}, \code{\link{compare1}}, \code{\link{compare2}}
}
